# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SwachhTrack Smart Waste Management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export interface SignupRequest {
  name: string;
  email: string;
  phone: string;
  password: string;
  city: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface UserProfile {
  id: string;
  name: string;
  email: string;
  phone: string;
  city: string;
  greenPoints: number;
  level: string;
  joinedAt: string;
}

export interface AuthResponse {
  token: string;
  user: UserProfile;
}

export interface ClassifyWasteRequest {
  imageBase64: string;
  userId?: string;
}

export type ClassificationResultCategory =
  (typeof ClassificationResultCategory)[keyof typeof ClassificationResultCategory];

export const ClassificationResultCategory = {
  organic: "organic",
  plastic: "plastic",
  metal: "metal",
  glass: "glass",
  ewaste: "ewaste",
  paper: "paper",
  other: "other",
} as const;

export interface ClassificationResult {
  sessionId: string;
  category: ClassificationResultCategory;
  itemName: string;
  confidence: number;
  disposalInstructions: string;
  binColor: string;
  tips: string[];
}

export interface VerifyDisposalRequest {
  sessionId: string;
  imageBase64: string;
  userId?: string;
  latitude?: number;
  longitude?: number;
}

export interface VerificationResult {
  verified: boolean;
  pointsAwarded: number;
  totalPoints: number;
  message: string;
  badge?: string | null;
}

export interface WasteEntry {
  id: string;
  userId: string;
  category: string;
  itemName: string;
  pointsEarned: number;
  verifiedAt: string;
  date: string;
}

export interface RewardInfo {
  userId: string;
  greenPoints: number;
  level: string;
  nextLevelPoints: number;
  rank: number;
  totalUsers: number;
  weeklyPoints: number;
  monthlyPoints: number;
}

export interface LeaderboardEntry {
  rank: number;
  userId: string;
  name: string;
  city: string;
  greenPoints: number;
  level: string;
  itemsRecycled: number;
}

export interface Badge {
  id: string;
  name: string;
  description: string;
  icon: string;
  earnedAt?: string | null;
  unlocked: boolean;
  requirement: string;
}

export interface WasteReportRequest {
  userId?: string;
  location: string;
  description: string;
  imageBase64?: string | null;
  latitude?: number | null;
  longitude?: number | null;
}

export type WasteReportStatus =
  (typeof WasteReportStatus)[keyof typeof WasteReportStatus];

export const WasteReportStatus = {
  pending: "pending",
  in_progress: "in_progress",
  resolved: "resolved",
} as const;

export interface WasteReport {
  id: string;
  userId?: string | null;
  location: string;
  description: string;
  status: WasteReportStatus;
  submittedAt: string;
  resolvedAt?: string | null;
}

export type AreaWiseDataCompliance =
  (typeof AreaWiseDataCompliance)[keyof typeof AreaWiseDataCompliance];

export const AreaWiseDataCompliance = {
  high: "high",
  medium: "medium",
  low: "low",
} as const;

export interface AreaWiseData {
  area: string;
  segregationRate: number;
  totalWaste: number;
  compliance: AreaWiseDataCompliance;
}

export interface WasteTypeData {
  category: string;
  percentage: number;
  amount: number;
}

export interface WeeklyTrend {
  day: string;
  organic: number;
  plastic: number;
  metal: number;
  glass: number;
  ewaste: number;
}

export interface CommunityData {
  name: string;
  area: string;
  score: number;
  members: number;
  rank: number;
}

export interface MunicipalAnalytics {
  totalWasteCollected: number;
  segregationRate: number;
  activeUsers: number;
  totalReports: number;
  areaWiseData: AreaWiseData[];
  wasteTypeDistribution: WasteTypeData[];
  weeklyTrend: WeeklyTrend[];
  topCommunities: CommunityData[];
}

export interface DailyActivity {
  date: string;
  points: number;
  items: number;
}

export interface CategoryCount {
  category: string;
  count: number;
  color: string;
}

export interface CitizenDashboard {
  totalPoints: number;
  itemsSegregated: number;
  ecoScore: number;
  rank: number;
  weeklyActivity: DailyActivity[];
  categoryBreakdown: CategoryCount[];
  recentActivity: WasteEntry[];
}

export interface PickupSchedule {
  id: string;
  wasteType: string;
  day: string;
  time: string;
  frequency: string;
  areas: string[];
  color: string;
  icon: string;
  notes?: string | null;
}

export interface RecyclingCenter {
  id: string;
  name: string;
  address: string;
  city: string;
  latitude: number;
  longitude: number;
  acceptedTypes: string[];
  timings: string;
  phone: string;
  distance?: string | null;
}

export type EcoChallengeDifficulty =
  (typeof EcoChallengeDifficulty)[keyof typeof EcoChallengeDifficulty];

export const EcoChallengeDifficulty = {
  easy: "easy",
  medium: "medium",
  hard: "hard",
} as const;

export interface EcoChallenge {
  id: string;
  title: string;
  description: string;
  points: number;
  category: string;
  difficulty: EcoChallengeDifficulty;
  completed: boolean;
  deadline: string;
}

export interface CompleteChallengeRequest {
  challengeId: string;
  userId?: string;
}

export interface CompleteChallengeResponse {
  success: boolean;
  pointsAwarded: number;
  message: string;
}

export interface EnvironmentalImpact {
  wasteDiverted: number;
  plasticRecycled: number;
  co2Reduced: number;
  treesEquivalent: number;
  waterSaved: number;
  energySaved: number;
}

export type GetWasteHistoryParams = {
  userId?: string;
};

export type GetRewardPointsParams = {
  userId?: string;
};

export type GetBadgesParams = {
  userId?: string;
};

export type GetCitizenDashboardParams = {
  userId?: string;
};

export type GetPickupScheduleParams = {
  city?: string;
};

export type GetRecyclingCentersParams = {
  city?: string;
};

export type GetEnvironmentalImpactParams = {
  userId?: string;
};