# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SwachhTrack Smart Waste Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/signup:
    post:
      operationId: signup
      tags: [auth]
      summary: User signup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: User login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /waste/classify:
    post:
      operationId: classifyWaste
      tags: [waste]
      summary: Classify waste from uploaded image
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClassifyWasteRequest"
      responses:
        "200":
          description: Classification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClassificationResult"
  /waste/verify:
    post:
      operationId: verifyWasteDisposal
      tags: [waste]
      summary: Verify waste disposal with second image
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyDisposalRequest"
      responses:
        "200":
          description: Verification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerificationResult"
  /waste/history:
    get:
      operationId: getWasteHistory
      tags: [waste]
      summary: Get user waste disposal history
      parameters:
        - in: query
          name: userId
          schema:
            type: string
          required: false
      responses:
        "200":
          description: Waste history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WasteEntry"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error, message]
    SignupRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        password:
          type: string
        city:
          type: string
      required: [name, email, phone, password, city]
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/UserProfile"
      required: [token, user]
    UserProfile:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        city:
          type: string
        greenPoints:
          type: number
        level:
          type: string
        joinedAt:
          type: string
      required: [id, name, email, phone, city, greenPoints, level, joinedAt]
    ClassifyWasteRequest:
      type: object
      properties:
        imageBase64:
          type: string
        userId:
          type: string
      required: [imageBase64]
    ClassificationResult:
      type: object
      properties:
        sessionId:
          type: string
        category:
          type: string
          enum: [organic, plastic, metal, glass, ewaste, paper, other]
        itemName:
          type: string
        confidence:
          type: number
        disposalInstructions:
          type: string
        binColor:
          type: string
        tips:
          type: array
          items:
            type: string
      required: [sessionId, category, itemName, confidence, disposalInstructions, binColor, tips]
    VerifyDisposalRequest:
      type: object
      properties:
        sessionId:
          type: string
        imageBase64:
          type: string
        userId:
          type: string
        latitude:
          type: number
        longitude:
          type: number
      required: [sessionId, imageBase64]
    VerificationResult:
      type: object
      properties:
        verified:
          type: boolean
        pointsAwarded:
          type: number
        totalPoints:
          type: number
        message:
          type: string
        badge:
          type: string
          nullable: true
      required: [verified, pointsAwarded, totalPoints, message]
    WasteEntry:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        category:
          type: string
        itemName:
          type: string
        pointsEarned:
          type: number
        verifiedAt:
          type: string
        date:
          type: string
      required: [id, userId, category, itemName, pointsEarned, verifiedAt, date]
...