# lib/api-zod/src/generated/types/verifyDisposalRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SwachhTrack Smart Waste Management API
 * OpenAPI spec version: 0.1.0
 */

export interface VerifyDisposalRequest {
  sessionId: string;
  imageBase64: string;
  userId?: string;
  latitude?: number;
  longitude?: number;
}